added suggested corrections

+++ db/migrate/20240722101043_create_debts.rb
@@ -1,7 +1,7 @@
 class CreateDebts < ActiveRecord::Migration[7.1]
   def change
     create_table :debts do |t|
-      t.decimal :amount, default: 0.0
+      t.float :amount, default: 0.0
       t.integer :from_user_id, null: false
       t.integer :to_user_id, null: false
       t.boolean :settled, default: false
